Improve DocC
