+++ Axys/Models/Axis/SegmentedMode.swift
@@ -5,6 +5,9 @@
 //  Created by Alessandro Bortoluzzi on 04/06/25.
 //
 
+/// Represents the mode of the inspector view.
+///
+/// Used to switch between manipulating an entity’s position or rotation.
 enum SegmentedMode {
     case position, rotation
 }
