+++ Axys/Models/Components/NameComponent.swift
@@ -7,6 +7,7 @@
 
 import RealityKit
 
+/// A component that stores the name of an object.
 struct NameComponent: Component {
 	var objectName: String
 }
